pyproject.toml
[project]
name = "stratum"
version = "0.1.0"
description = "CrewAI-based multi-agent system for recursive forensic audits of scientific papers"
requires-python = ">=3.10,<3.14"
authors = [
    {name = "Stratum Team"}
]
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.20.0",
    "litellm>=1.50.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pymupdf>=1.24.0",
    "grobid-tei-xml>=0.1.3",
    "requests>=2.32.0",
    "pyyaml>=6.0",
    "typer>=0.15.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
]

[project.scripts]
stratum = "stratum.main:app"

[build-system]
requires = ["setuptools>=75.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.crewai]
type = "flow"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=src/stratum --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
